Move max height detection to use-popper and fix types
